Revert "Revert "Implement lead report page""
